+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DO193vVa.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-rvxnGZdt.js";import{_ as a}from"./mermaid.core-D1GPJ5P9.js";import"./chunk-5VM5RSS4-8T4j41UN.js";import"./chunk-XXDRQBXY-CVBuJc7z.js";import"./chunk-POPQ4Y6H-CIPzuB3G.js";import"./chunk-F27PBJKO-DofrUHWv.js";import"./channel-B9u-1QiG.js";import"./index-Bq68lW3g.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};